django_api/requirements.txt
django==3.2
flake8==4.0
python-dotenv==0.21.0
psycopg2-binary==2.9.5
django-split-settings==1.2.0
django-debug-toolbar==3.4.0

